requirements.txt
appnope==0.1.0
backports-abc==0.4
backports.lzma==0.0.6
backports.shutil-get-terminal-size==1.0.0
bitarray==0.8.1
bitstring==3.1.5
Brotli==0.5.2
cairocffi==0.7.2
certifi==2016.9.26
cffi==1.8.3
configparser==3.5.0
contextlib2==0.5.4
cycler==0.10.0
dask==0.12.0
decorator==4.0.10
entrypoints==0.2.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
gizeh==0.1.10
Glymur==0.8.7
h5py==2.6.0
imageio==1.6
imread==0.6
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
llvmlite==0.14.0
lzw==0.1.11
mahotas==1.4.3
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
moviepy==0.2.2.11
nbconvert==4.2.0
nbformat==4.1.0
networkx==1.11
notebook==4.2.3
numba==0.29.0
numpy==1.11.2
pandas==0.19.1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
ptyprocess==0.5.1
purepng==0.2.0
pycparser==2.17
Pygments==2.1.3
pyparsing==2.1.10
python-dateutil==2.5.3
python-lzf==0.2.4
python-lzo==1.11
pytz==2016.7
pyzmq==16.0.0
qtconsole==4.2.1
scikit-image==0.12.3
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
toolz==0.8.1
tornado==4.4.2
tqdm==4.9.0
traitlets==4.3.1
wcwidth==0.1.7
wheel==0.24.0
widgetsnbextension==1.2.6
zopfli==0.0.3
zstd==1.0.0.99.1
